refactor: Flatten esa directory
